% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workshop_certificate.R
\name{create_workshop_certificates}
\alias{create_workshop_certificates}
\title{Create certificates for all attendees}
\usage{
create_workshop_certificates(
  attendees,
  workshop,
  date,
  location,
  curriculum,
  certifier,
  credentials,
  organization,
  organization_url,
  dir = ".",
  title = "CERTIFICATE OF COMPLETION",
  action_text = "participated in the",
  curriculum_title = "Workshop contents:",
  signature = NULL,
  signature_skip = -0.8,
  logo = "R",
  border_image = NULL,
  papersize = "a4paper",
  keep_tex = FALSE
)
}
\arguments{
\item{attendees}{Names of attendees, character vector.}

\item{workshop}{Workshop title, character.}

\item{date}{Date of the workshop, as a date.}

\item{location}{Location of the workshop, character.}

\item{curriculum}{Path to the workshop curriculum (.md), character.}

\item{certifier}{Person certifying, character.}

\item{credentials}{Credentials of the certifying person, character.}

\item{organization}{Description of the organization running the workshops.}

\item{organization_url}{URL of the organization website, without https://.}

\item{dir}{Directory where to output the pdf certificates, character.}

\item{title}{Title for certificate.}

\item{action_text}{Action text between name and workshop title.}

\item{curriculum_title}{Header before curriculum content.}

\item{signature}{The path to an image to use as a signature.}

\item{signature_skip}{The space to skip between the signature image and the
signature line (units: cm).}

\item{logo}{The logo to use for the watermark, as a path to a logo
file or the name of a logo included in the package: either `"R"`
(default), or `"Forwards"`.}

\item{border_image}{Specify an image to use to create a border, as a path
to an image file. Defaults are used for the included logos: plain blue for
`logo = R` and a magma colour scale for `logo = "Forwards"`.}

\item{papersize}{Option for LaTeX article class specifying paper size, e.g.
`"a4paper"`, `"letterpaper"`.}

\item{keep_tex}{Logical argument passed to rmarkdown::render.}
}
\description{
Create certificates for all attendees
}
\examples{
\dontrun{
# Fake names generated via charlatan::ch_name
attendees <- c("Marnie Dickinson", "Dr. Marlin Wilderman")
workshop <- "Package development workshop"
date <- as.Date("2018-01-01")
location <- "University of Lorraine"
curriculum <- system.file("rmarkdown", "templates",
"workshop_certificate", "resources",
"default_workshop_contents.md", package = "certificate")
certifier <- "Zaire Crooks"
credentials <- "Forwards teaching team member"
organization <- "Forwards, the R Foundation taskforce for women and other
under-represented groups"
organization_url <- "forwards.github.io/"
dir <- "certificates"
create_workshop_certificates(attendees, workshop, date, location,
                             curriculum, certifier, credentials,
                             organization, organization_url,
                             dir)
}
}
