% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workshop_certificate.R
\name{create_workshop_certificates}
\alias{create_workshop_certificates}
\title{Create certificates for all attendees}
\usage{
create_workshop_certificates(date, location, workshop, curriculum,
  certifier, credentials, attendees, dir = ".", papersize = "a4paper",
  keep_tex = FALSE)
}
\arguments{
\item{date}{Date of the workshop, as a date.}

\item{location}{Location of the workshop, character.}

\item{workshop}{Workshop title, character.}

\item{curriculum}{Path to the workshop curriculum (.md), character.}

\item{certifier}{Person certifying, character.}

\item{credentials}{Credentials of the certifying person, character.}

\item{attendees}{Names of attendees, character vector.}

\item{dir}{Directory where to output the pdf certificates, character.}

\item{papersize}{Option for LaTeX article class specifying paper size, e.g.
`"a4paper"`, `"letterpaper"`.}

\item{keep_tex}{Logical argument passed to rmarkdown::render}
}
\description{
Create certificates for all attendees
}
\examples{
\dontrun{
# Fake names generated via charlatan::ch_name
attendees <- c("Marnie Dickinson", "Dr. Marlin Wilderman")
date <- as.Date("2018-01-01")
location <- "University of Lorraine"
workshop <- "Package development workshop"
curriculum <- system.file("rmarkdown", "templates",
"workshop_certificate", "resources",
"default_workshop_contents.md", package = "fwdbrand")
certifier <- "Zaire Crooks"
credentials <- "Forwards teaching team member"
dir <- "certificates"
create_workshop_certificates(date, location, workshop, curriculum,
 certifier,
credentials,
attendees,
dir)
}
}
